kc-core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kc-core"
version = "0.1.0"
description = "KonaClaw core agent runtime"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
    "ruff>=0.4",
]

[project.scripts]
kc-chat = "kc_core.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/kc_core"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v"

[tool.ruff]
line-length = 100
target-version = "py311"
